Bench 1756172

+++ Schoenemann/tt.h
@@ -8,7 +8,7 @@ using namespace chess;
 const short EXACT = 0;
 const short UPPER_BOUND = 1;
 const short LOWER_BOUND = 2; //Lower bound
-const short infinity = 32767;
+const int infinity = 32767;
 
 struct Hash {
     std::uint64_t key;
